ai-ml/bedrock-nova-image-prompt-optimizer/requirements.txt
streamlit
pillow
boto3
